Add compatibility ZSH prompts and color utilities
There are two versions of the new prompt. One that still uses ANSI and
one that works without color. Both do not depend on Powerline Fonts
anymore.
There are now commands for changing the monitor brightness and gamma
values. Due to bugs with Gnome Night Light and not being able to get the
same look, the commands are still rather rough.